+++ Vector.h
@@ -4,16 +4,16 @@
 #include <string>
 
 class Vector {
-	double sklX; //!< skladowa x wektora
-	double sklY; //!< skladowa y wektora
-	double sklZ; //!< skladowa z wektora
+	long double sklX; //!< skladowa x wektora
+	long double sklY; //!< skladowa y wektora
+	long double sklZ; //!< skladowa z wektora
 
 public:
 	/**
 	 * @brief Konstruktor tworzacy instancje klasy Vector z dwoma argumentami
 	 * @arg a, b to poczatkowe skladowe wektora
 	 */
-	Vector(double a, double b, double c);
+	Vector(long double a, long double b, long double c);
 	/**
 	 * @brief Konstruktor tworzacy instancje klasy Vector ze skladowymi ustawianymi na 0.
 	 */
@@ -22,32 +22,32 @@ public:
 	 * @brief Funkcja ustawiajaca skladowa x wektora
 	 * @arg a to ustawiana skladowa x wektora
 	 */
-	void setX(double a);
+	void setX(long double a);
 	/**
 	 * @brief Funkcja ustawiajaca skladowa y wektora
 	 * @arg b to ustawiana skladowa y wektora
 	 */
-	void setY(double b);
+	void setY(long double b);
 	/**
 	 * @brief Funkcja ustawiajaca skladowa z wektora
 	 * @arg b to ustawiana skladowa z wektora
 	 */
-	void setZ(double c);
+	void setZ(long double c);
 	/**
 	 * @brief Funkcja zwracajaca skladowa x wektora
 	 * @return skladowa x wektora
 	 */
-	double x();
+	long double x();
 	/**
 	 * @brief Funkcja zwracajaca skladowa y wektora
 	 * @return skladowa y wektora
 	 */
-	double y();
+	long double y();
 	/**
 	 * @brief Funkcja zwracajaca skladowa z wektora
 	 * @return skladowa z wektora
 	 */
-	double z();
+	long double z();
 
 	std::string toString();
 };
